backend-fastapi/pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "BIT-MITRA FastAPI Backend with async SQLAlchemy"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "pyjwt>=2.8.0",
    "pwdlib[argon2]>=0.2.0",
    "python-multipart>=0.0.6",
    "cloudinary>=1.38.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
    "pillow>=10.2.0",
    "email-validator>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
